pyproject.toml
[project]
name = "prueba-tecnica-ia"
version = "0.1.0"
description = "APIs FastAPI con JWT, cargas a S3, SQL Server y analítica asistida por IA."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.0.3",
    "sqlalchemy>=2.0.20",
    "pyodbc>=5.1.0",
    "boto3>=1.34.0",
    "transformers>=4.37.0",
    "torch>=2.2.0",
    "sentencepiece>=0.1.99"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.13"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

